chore: fix demo so it runs with ts-node
